+++ CHANGELOG.md
@@ -121,6 +121,14 @@
     * Agnidus agate sliver
   * Fixed Itto rarity
   * Changed links for embedded map again because of it's update
+  * Added farming routes for:
+    * Nilotpala Lotus
+    * Rukkhashava Mushrooms
+    * Kalpalata Lotus
+    * Faded/Trimmed/Rich red satin
+    * Fungal spores
+    * Luminescent pollen
+    * Crystalline cyst dust
 
 #### 1.1.9 June-2022